+++ library/src/device/generator/stockham_gen_rr.h
@@ -25,7 +25,7 @@ struct StockhamKernelRR : public StockhamKernel
     explicit StockhamKernelRR(StockhamGeneratorSpecs& specs)
         : StockhamKernel(specs)
     {
-        load_from_lds = false;
+        direct_to_reg = true;
     }
 
     std::string tiling_name() override