+++ CHANGELOG.md
@@ -9,8 +9,12 @@ Full documentation for rocFFT is available at [rocfft.readthedocs.io](https://ro
 
 ### Changed
 - Improved reuse of twiddle memory between plans.
-- Set a default load/store callback when only one callback 
+- Set a default load/store callback when only one callback
   type is set via the API for improved performance.
+
+### Optimizations
+- Introduced a new access pattern of lds (non-linear) and applied it on
+  sbcc kernels len 64 and 81 to get performance improvement.
 
 ## (Unreleased) rocFFT 1.0.16